Optimize

+++ src/routes/RestaurantListPage/RestaurantListPage.js
@@ -3,6 +3,7 @@ import WishlistItem from '../../components/WishlistItem/WishlistItem';
 import VisitedItem from '../../components/VisitedItem/VisitedItem';
 import { Link } from 'react-router-dom';
 import RestaurantsApiService from '../../services/restaurant-api-service';
+import Swal from 'sweetalert2';
 import './RestaurantListPage.css';
 
 
@@ -87,10 +88,23 @@ export default class RestaurantListPage extends Component {
     }
 
     onDelete = (id) => {
-        RestaurantsApiService.deleteUserRestaurant(id)
-        let newRestaurants = this.state.restaurantsList.filter(restaurant => restaurant.id !== id)
-        this.setState({
-            restaurantsList: newRestaurants
+        let restaurant = this.state.restaurantsList.find(itm => itm.id === id)
+        Swal.fire({
+            title: 'Are you sure?',
+            text: `You are about to delete ${restaurant.restaurant.name}.`,
+            icon: 'warning',
+            confirmButtonText: "I'm sure!",
+            showCancelButton: true,
+        })
+        .then(value => {
+            if (value.dismiss === 'cancel') return null
+            else {
+                RestaurantsApiService.deleteUserRestaurant(id)
+                let newRestaurants = this.state.restaurantsList.filter(restaurant => restaurant.id !== id)
+                this.setState({
+                    restaurantsList: newRestaurants
+                })
+            }
         })
     }
 
@@ -130,14 +144,21 @@ export default class RestaurantListPage extends Component {
     renderFilterWishlist() {
         return (
             <div className="wishlist_filter">
-                <label htmlFor="city_filter">Filter by city: </label>
-                <input type="test" id="city_filter" value={this.state.cityFilter} onChange={this.handleCityChange}></input>
-                <label htmlFor="cuisine_filter">Filter by cuisine: </label>
-                <select id="cuisine_filter" onChange={this.handleCuisineChange}>
-                    {this.state.cuisines.map(cuisine => {
-                        return <option key={cuisine.id}>{cuisine.cuisine_name}</option>
-                    })}
-                </select>
+                <div className="wishlist_filter">
+                    <label className="filter_label" htmlFor="city_filter">Filter by city: 
+                        <input type="test" id="city_filter" value={this.state.cityFilter} onChange={this.handleCityChange}></input>
+                    </label>
+                </div>
+                <div className="wishlist_filter">
+                    <label className="filter_label" htmlFor="cuisine_filter">Filter by cuisine: 
+                    <select id="cuisine_filter" onChange={this.handleCuisineChange}>
+                        {this.state.cuisines.map(cuisine => {
+                            return <option key={cuisine.id}>{cuisine.cuisine_name}</option>
+                        })}
+                    </select>
+                    </label>
+                </div>
+                <button id="filter" onClick={this.toggleWishListFilter}>Cancel</button>
             </div>
         )
     }
@@ -145,22 +166,32 @@ export default class RestaurantListPage extends Component {
     renderFilterVisited() {
         return (
             <div className="visited_filter">
-                <label htmlFor="city_filter">Filter by city: </label>
-                <input type="test" id="city_filter" value={this.state.cityFilter} onChange={this.handleCityChange}></input>
-                <label htmlFor="cuisine_filter">Filter by cuisine: </label>
-                <select id="cuisine_filter" onChange={this.handleCuisineChange}>
-                    {this.state.cuisines.map(cuisine => {
-                        return <option key={cuisine.id}>{cuisine.cuisine_name}</option>
-                    })}
-                </select>
-                <label htmlFor="rating_filter">Filter by rating: </label>
-                <select id="rating_filter" onChange={this.handleRatingChange}>
-                    <option>5 stars</option>
-                    <option>4 stars and up</option>
-                    <option>3 stars and up</option>
-                    <option>2 stars and up</option>
-                    <option>1 star and up</option>
-                </select>
+                <div className="visited_filter">
+                    <label className="filter_label" htmlFor="city_filter">Filter by city: 
+                        <input type="test" id="city_filter" value={this.state.cityFilter} onChange={this.handleCityChange}></input>
+                    </label>
+                </div>
+                <div className="visited_filter">
+                    <label className="filter_label" htmlFor="cuisine_filter">Filter by cuisine: 
+                    <select id="cuisine_filter" onChange={this.handleCuisineChange}>
+                        {this.state.cuisines.map(cuisine => {
+                            return <option key={cuisine.id}>{cuisine.cuisine_name}</option>
+                        })}
+                    </select>
+                    </label>
+                </div>
+                <div className="visited_filter">
+                    <label className="filter_label" htmlFor="rating_filter">Filter by rating: 
+                    <select id="rating_filter" onChange={this.handleRatingChange}>
+                        <option>5 stars</option>
+                        <option>4 stars and up</option>
+                        <option>3 stars and up</option>
+                        <option>2 stars and up</option>
+                        <option>1 star and up</option>
+                    </select>
+                    </label>
+                </div>
+                <button id="filter" onClick={this.toggleVisitedFilter}>Cancel</button>
             </div>
         )
     }
@@ -205,7 +236,11 @@ export default class RestaurantListPage extends Component {
             restaurants = restaurants.filter(itm => itm.restaurant.cuisine_name.toLowerCase() === filter)
         }
         if (!restaurants.length) {
-            return <span>No results</span>
+            if (this.state.wishlistFilter) {
+                return <p className="empty_text">No Results</p>
+            } else {
+                return <p className="empty_text">You don't have any wishlist restaurants yet! To start adding, click the "Add Restaurant" button above.</p>
+            }
         }
         return restaurants.map(restaurant => <WishlistItem key={restaurant.id} restaurant={restaurant} onDelete={this.onDelete} onMove={this.onMove} />)
     }
@@ -225,7 +260,11 @@ export default class RestaurantListPage extends Component {
             restaurants = restaurants.filter(itm => itm.rating >= filter)
         }
         if (!restaurants.length) {
-            return <span>No results</span>
+            if (this.state.visitedFilter) {
+                return <p className="empty_text">No Results</p>
+            } else {
+                return <p className="empty_text">You don't have any visited restaurants yet! To start adding, click the "Add Restaurant" button above or mark one of your wishlist restaurants as visited.</p>
+            }
         }
         return restaurants.map(restaurant => <VisitedItem key={restaurant.id} restaurant={restaurant} onDelete={this.onDelete} onEdit={this.onEdit} />)
     }
@@ -233,7 +272,7 @@ export default class RestaurantListPage extends Component {
     renderWishlist(error) {
         return (
             <>
-                <button type="button" className="filter" onClick={this.toggleWishListFilter}>{this.state.wishlistFilter ? 'Cancel Filter' : 'Filter by'}</button>
+                <button type="button" id="filter" onClick={this.toggleWishListFilter}>Filter</button>
                 {this.state.wishlistFilter ? this.renderFilterWishlist() : ''}
                 <ul>
                     {error ? <p className='red'>There was an error, try again</p> : this.renderWishlistRestaurants()}
@@ -245,7 +284,7 @@ export default class RestaurantListPage extends Component {
     renderVisited(error) {
         return (
             <>
-                <button type="button" className="filter" onClick={this.toggleVisitedFilter}>{this.state.visitedFilter ? 'Cancel Filter' : 'Filter by'}</button>
+                <button type="button" id="filter" onClick={this.toggleVisitedFilter}>Filter</button>
                 {this.state.visitedFilter ? this.renderFilterVisited() : ''}
                 <ul>
                     {error ? <p className='red'>There was an error, try again</p> : this.renderVisitedRestaurants()}
